Extract a request appointment time component and add to group items
