+++ app/components/aeon/request_appointment_time_component.html.erb
@@ -1,0 +1,5 @@
+<span class="appointment-details ms-1">
+  <span class="fw-bold text-green"><%= appointment_date %></span>
+  <span class="fw-normal text-nowrap text-green"><%= appointment_time_range %></span>
+  <% if @with_reading_room %><span class="fw-normal text-nowrap text-green"><%= appointment_reading_room_name %></span><% end %>
+</span>
