+++ app/components/aeon/request_group_item_component.html.erb
@@ -4,7 +4,10 @@
       <%= render Aeon::RequestItemIdentifierComponent.new(request:) %>
       <%= render Aeon::RequestItemDetailComponent.new(request:) %>
     </div>
-    <%= render Aeon::RequestActionsComponent.new(request:) %>
+    <div class="d-flex align-items-center gap-5 justify-content-between">
+      <%= render Aeon::RequestAppointmentTimeComponent.new(request:, with_reading_room: false) %>
+      <%= render Aeon::RequestActionsComponent.new(request:) %>
+    </div>
   </div>
   <div class="pt-2 fs-14 d-flex gap-sm-3 flex-sm-row flex-column">
     <div>Request #<%= transaction_number %></div>
